requirements.txt
streamlit
pyodbc
google-generativeai
